Refactor project structure and fix audio loading
This commit refactors the project by separating the core logic into modules for audio processing and file management. It also resolves a critical bug where saved audios and their transcriptions were not loading correctly in the UI.

Key changes:
- Created `src/audio_processing.py` to house the audio transcription and diarization logic.
- Created `src/file_management.py` to manage all interactions with the audio library, including saving, loading, and metadata handling.
- Refactored `app.py` to focus on the Gradio UI, importing logic from the new modules.
- Fixed the event handler for the library browser to correctly load selected audio files and display their saved transcriptions without requiring reprocessing.
